Add resnet152 Fit_Generator.ipynb
Add the notebook where I trained the resnet152 model using the fit_generator(). From the documentation I've read, I should've used the fit() to train it. Perhaps that was the issue with the other lung cancer models bar the resnet50, as that was trained using fit() and showed improvement over time in loss and accuracy while the others were stagnant and didn't improve in accuracy.