Add ResNet50 Kaggle Notebook run
I was able to modify the notebook to work with the Kaggle dataset. The results are a bit all over the place, I would need to add an monitor and early stopping to stabilize the training. I trained it using the GPU T4 x2 accelerator and it took about 9 hours to complete. I am finally able to get some model training right, I should move onto vision transformers since there is a limit on the accelerator usage.