Add Resnet152 Kaggle Runs
The key differences in the model from other runs is that the model was too big to run on the target size of (256,256), so it had to be shrunken down to (128,128). I also changed the last output dimension from 2 to 1 and changed the activation to sigmoid from softmax for the resnet152-layer-change. I seemed to me that this would work better, at least logically and through resources discussing binary classification. The final training and validation accuracy seemed good for the resnet152-no-layer-change - it was all over the place, so making a small change could hopefully steady the curve-, but when evaluated on the test set it scored a 43.89%. As seen in resnet152-layer-change, the test accuracy improved to 49.33%, but the training and validation accuracy did not improve at all. It could potentially be due to Kaggle's GPUs being used up; I saw early on in training that the GPU wasn't being used constantly, probably due to an overload from the system.

I should move onto transformers. I've looked at VTs and CVTs, but don't quite understand them and how I can make the dataset best utilized for them.